requirements.txt
python-uinput>=0.11.2
cec>=0.2.6
